requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
jinja2==3.1.6
python-multipart==0.0.20
pydantic-settings==2.10.1
python-slugify==8.0.4
httpx==0.28.1
pytest==8.4.1
